# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EB Insights API",
    "version": "2.0.0",
    "description": "Backend REST API do EB Insights — sincronização multi-coletor de coletas de Escola Bíblica (Spec 007).\n\n**Fonte canônica**: este arquivo. Rationale por domínio: [`specs/007-sync-backend/contracts/`](../../specs/007-sync-backend/contracts/). Envelope de export v2: [`specs/005-export-contract-v2/contracts/export-envelope.v2.schema.json`](../../specs/005-export-contract-v2/contracts/export-envelope.v2.schema.json).\n\n**Convenções**:\n- Autenticação por `Authorization: Bearer <jwt>` (JWT HS256, TTL 7 dias).\n- Erros seguem `{ code, message }` com `code` em snake_case (registro em `specs/007-sync-backend/contracts/error-codes.md`). Mensagens em pt-BR.\n- Primeiro usuário registrado é auto-promovido a `COORDINATOR`; demais entram como `COLLECTOR`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Dev local (mesma máquina)"
    },
    {
      "url": "http://{lanIp}:3000",
      "description": "Dev local via LAN (Expo Go em celular físico)",
      "variables": {
        "lanIp": {
          "default": "192.168.0.42",
          "description": "IPv4 da máquina na LAN. No Windows, obter via `ipconfig`."
        }
      }
    }
  ],
...
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/catalog": {
      "get": {
        "tags": ["catalog"],
        "summary": "Catálogo (séries + tópicos + professores)",
        "description": "Retorna as 3 coleções de catálogo. `?since=` filtra por `updated_at > since` (leitura incremental). `?include_pending=true` inclui itens criados automaticamente via `/sync/batch` e ainda não moderados — só coordinator tem permissão.",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "ISO-8601. Filtra por `updated_at > since`.",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "include_pending",
            "in": "query",
            "required": false,
            "description": "`true` inclui itens `is_pending = true`. Requer `role = COORDINATOR`.",
            "schema": { "type": "string", "enum": ["true", "false"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo completo (ou filtrado).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CatalogResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" }
        }
      }
    },
    "/catalog/series": {
      "post": {
        "tags": ["catalog"],
        "summary": "Criar série",
        "description": "Cria uma nova série de lições. Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SeriesCreate" },
              "example": { "code": "EB400", "title": "Gálatas", "description": "Série de estudos em Gálatas" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Série criada.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SeriesItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      }
    },
    "/catalog/series/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "minLength": 1 }
        }
      ],
      "patch": {
        "tags": ["catalog"],
        "summary": "Atualizar série",
        "description": "Atualização parcial. Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SeriesPatch" },
              "example": { "title": "Gálatas — revisado" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Série atualizada.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SeriesItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      },
      "delete": {
        "tags": ["catalog"],
        "summary": "Excluir série",
        "description": "Só permitido se não houver tópicos vinculados (409 `series_referenced` caso contrário). Requer `role = COORDINATOR`.",
        "responses": {
          "204": { "description": "Série excluída." },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      }
    },
    "/catalog/topics": {
      "post": {
        "tags": ["catalog"],
        "summary": "Criar tópico",
        "description": "Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TopicCreate" },
              "example": {
                "series_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                "title": "Lição 1 — Introdução",
                "sequence_order": 1,
                "suggested_date": "2026-04-11"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tópico criado.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TopicItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" }
        }
      }
    },
    "/catalog/topics/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
      ],
      "patch": {
        "tags": ["catalog"],
        "summary": "Atualizar tópico",
        "description": "Atualização parcial. Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TopicPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tópico atualizado.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TopicItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" }
        }
      },
      "delete": {
        "tags": ["catalog"],
        "summary": "Excluir tópico",
        "description": "409 `topic_referenced` se há coletas apontando pra ele. Requer `role = COORDINATOR`.",
        "responses": {
          "204": { "description": "Tópico excluído." },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      }
    },
    "/catalog/professors": {
      "post": {
        "tags": ["catalog"],
        "summary": "Criar professor",
        "description": "Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfessorCreate" },
              "example": { "name": "João Silva", "email": "joao@example.com", "doc_id": "11144477735" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Professor criado.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfessorItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      }
    },
    "/catalog/professors/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1 } }
      ],
      "patch": {
        "tags": ["catalog"],
        "summary": "Atualizar professor",
        "description": "Atualização parcial. Requer `role = COORDINATOR`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfessorPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Professor atualizado.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ProfessorItem" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error400" },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      },
      "delete": {
        "tags": ["catalog"],
        "summary": "Excluir professor",
        "description": "409 `professor_referenced` se há coletas apontando pra ele. Requer `role = COORDINATOR`.",
        "responses": {
          "204": { "description": "Professor excluído." },
          "401": { "$ref": "#/components/responses/Error401" },
          "403": { "$ref": "#/components/responses/Error403" },
          "404": { "$ref": "#/components/responses/Error404" },
          "409": { "$ref": "#/components/responses/Error409" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT HS256 obtido via `POST /auth/register` ou `POST /auth/login`. Validade: 7 dias."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Código estável em snake_case. Registro completo: `specs/007-sync-backend/contracts/error-codes.md`.",
            "example": "unauthenticated"
          },
          "message": {
            "type": "string",
            "description": "Mensagem em pt-BR, segura pra exibir ao usuário.",
            "example": "Credencial ausente ou inválida."
          }
        }
      },
...
      "SeriesItem": {
        "type": "object",
        "required": ["id", "code", "title", "is_pending", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "code": { "type": "string", "description": "Código único (ex: EB400)." },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "is_pending": {
            "type": "boolean",
            "description": "`true` quando a série foi auto-criada via `/sync/batch` e aguarda moderação."
          },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "SeriesCreate": {
        "type": "object",
        "required": ["code", "title"],
        "properties": {
          "code": { "type": "string", "minLength": 1 },
          "title": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "nullable": true }
        }
      },
      "SeriesPatch": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "minLength": 1 },
          "title": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "nullable": true }
        }
      },
      "TopicItem": {
        "type": "object",
        "required": ["id", "series_id", "title", "sequence_order", "is_pending", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "series_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "sequence_order": { "type": "integer" },
          "suggested_date": { "type": "string", "format": "date", "nullable": true },
          "is_pending": { "type": "boolean" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "TopicCreate": {
        "type": "object",
        "required": ["series_id", "title", "sequence_order"],
        "properties": {
          "series_id": { "type": "string", "minLength": 1 },
          "title": { "type": "string", "minLength": 1 },
          "sequence_order": { "type": "integer" },
          "suggested_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "TopicPatch": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1 },
          "sequence_order": { "type": "integer" },
          "suggested_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "ProfessorItem": {
        "type": "object",
        "required": ["id", "doc_id", "name", "is_pending", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "doc_id": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "nullable": true,
            "description": "CPF (11 dígitos, sem máscara). `null` para registros legados anteriores ao campo."
          },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "is_pending": { "type": "boolean" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProfessorCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "id": { "type": "string", "minLength": 1 },
          "name": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email", "nullable": true },
          "doc_id": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "nullable": true,
            "description": "CPF (11 dígitos sem máscara)."
          }
        }
      },
      "ProfessorPatch": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "email": { "type": "string", "format": "email", "nullable": true },
          "doc_id": {
            "type": "string",
            "pattern": "^[0-9]{11}$",
            "nullable": true
          }
        }
      },
      "CatalogResponse": {
        "type": "object",
        "required": ["series", "topics", "professors", "server_now"],
        "properties": {
          "series": { "type": "array", "items": { "$ref": "#/components/schemas/SeriesItem" } },
          "topics": { "type": "array", "items": { "$ref": "#/components/schemas/TopicItem" } },
          "professors": { "type": "array", "items": { "$ref": "#/components/schemas/ProfessorItem" } },
          "server_now": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Bad request — payload ou query inválidos.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "invalid_payload": { "value": { "code": "invalid_payload", "message": "Payload inválido." } },
              "invalid_query": { "value": { "code": "invalid_query", "message": "Parâmetros de consulta inválidos." } },
              "password_too_short": { "value": { "code": "password_too_short", "message": "A senha deve ter no mínimo 8 caracteres." } },
              "invalid_email": { "value": { "code": "invalid_email", "message": "E-mail inválido." } }
            }
          }
        }
      },
      "Error401": {
        "description": "Não autenticado ou credenciais inválidas.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "unauthenticated": { "value": { "code": "unauthenticated", "message": "Credencial ausente ou inválida." } },
              "invalid_credentials": { "value": { "code": "invalid_credentials", "message": "E-mail ou senha inválidos." } }
            }
          }
        }
      },
      "Error403": {
        "description": "Autenticado mas sem permissão (role insuficiente).",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "code": "forbidden", "message": "Acesso restrito a coordenadores." }
          }
        }
      },
      "Error404": {
        "description": "Recurso não encontrado.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": { "code": "not_found", "message": "Registro não encontrado." }
          }
        }
      },
      "Error409": {
        "description": "Conflito — duplicado ou com referências.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "examples": {
              "email_already_registered": { "value": { "code": "email_already_registered", "message": "Este e-mail já está cadastrado." } },
              "code_already_exists": { "value": { "code": "code_already_exists", "message": "Código já existe." } },
              "series_referenced": { "value": { "code": "series_referenced", "message": "Série não pode ser excluída enquanto houver aulas associadas." } },
              "topic_referenced": { "value": { "code": "topic_referenced", "message": "Tópico não pode ser excluído enquanto houver aulas associadas." } },
              "professor_referenced": { "value": { "code": "professor_referenced", "message": "Professor não pode ser excluído enquanto houver aulas associadas." } }
            }
          }
        }
      },
...
    }
  }
}